feat: add CustomerDeliveryDetailViewAllContainer
