+++ packages/apps/stock/src/components/templates/index.js
@@ -19,6 +19,7 @@
 export {default as CustomerDeliveryCard} from './CustomerDeliveryCard/CustomerDeliveryCard';
 export {default as CustomerDeliveryDetailHeader} from './CustomerDeliveryDetailHeader/CustomerDeliveryDetailHeader';
 export {default as CustomerDeliveryDetailMovementIndicationCard} from './CustomerDeliveryDetailMovementIndicationCard/CustomerDeliveryDetailMovementIndicationCard';
+export {default as CustomerDeliveryDetailViewAllContainer} from './CustomerDeliveryDetailViewAllContainer/CustomerDeliveryDetailViewAllContainer';
 export {default as CustomerDeliveryLineCard} from './CustomerDeliveryLineCard/CustomerDeliveryLineCard';
 export {default as CustomerDeliveryLineDetailFixedItems} from './CustomerDeliveryLineDetailFixedItems/CustomerDeliveryLineDetailFixedItems';
 export {default as CustomerDeliveryLineQuantityCard} from './CustomerDeliveryLineQuantityCard/CustomerDeliveryLineQuantityCard';
